find-park-form no params

+++ ParkIT/src/main/java/es/ucm/fdi/iw/controller/UserController.java
@@ -97,7 +97,7 @@ public class UserController {
 	@GetMapping("/map")
 	public String map(@RequestParam @Nullable String address, @RequestParam @Nullable String startDate, @RequestParam @Nullable String endDate, 
 		@RequestParam @Nullable String startTime, @RequestParam @Nullable String endTime, @RequestParam @Nullable String latitude,
-		@RequestParam @Nullable String longitude, @RequestParam @Nullable Double range, Model model) {
+		@RequestParam @Nullable String longitude, @RequestParam @Nullable Long range, Model model) {
 
 		System.out.println("Address: " + address);
 		List<Parking> parkings;
